Single replacement pass through the template
Currently we do three passes over the template, one for each type of replacement. This means that we can start replacing existing replacements (as would be the case with handlebars). This merges the replacement regexps into a single pattern and runs a single pass over the template.

The actual replacement code is pretty ugly now. Would be good to refactor at some point to tidy it up a bit.