Add an expectation of node version to package.json
I don't think this is automatically enforced, or automatically switched
too (like rvm or such), but it gives a hint that older versions of node
won't work with this package.

I picked 6 as something reasonably modern, within the last 6 months, but
it's a bit arbitrary. I know i had problems with a version <= 5 though.