Stop serving separate font CSS file
It was already bundled with main.scss, so the separate file was duplicated CSS.

The two font stylesheet includes were removed from the layout template, to avoid
the duplication, but this does remove IE8 font support (though it didn't actually
work as there isn't a IE8 font CSS file in the alpha), which we'll need to decide
how to support before the beta. I've captured that in this issue:

https://github.com/alphagov/govuk_frontend_alpha/issues/58

Also, treat the font CSS as a SASS module, rather than a top level/manifest file,
as thats how we're using it.