Use BEM for error states
Ensure that —has-error is used as a modifier.
This is to defend against the likely event that others are already
using .has-error in their codebase.

Ideally, we’d use .has-error as a state based class to chain to the
element with the error and ensure that bare state classes aren’t used.