Add browser-sync and a set of watch tasks
Use the watch tasks to watch for changes to the scss, js and img
folders in app/assets and call the build tasks.

Amend the config file to rename paths.nodeModules, update anywhere this
was being used as paths.node_modules.

Set nodemon to ignore changes to the public,
dist and node_modules folder.