Add rails components helper
It would be nicer if this lived in `app/helpers` and got autoloaded, but
I cannot get that to work, so don't care for the alpha.

To include in a consuming app, add `helper GovukFrontendAlpha::ApplicationHelper`
 to ApplicationController, and invoke in a view like:

 ```
 <%= govuk_component("button", {
  text: 'My Button',
}) %>
```

Again, it would be nice if it was automatically loaded as part of the engine,
but not sure how, and not essential for the alpha.