Generate nunjuck components/macros in build step
Macros are a pretty nunjucks/jinja specific structure, which don't have
equivilents in all the languages we want to support. It also doesn't work
well with Fractal, which expects non macro component templates.

Treat the macro syntax as something package-specific for nunjucks output,
and drop it from the source definition and move the arguments to the
component config, for use in packaging and potentially documentation.

The config structure for arguments is just the names of variables, so
won't support default arguments, or documentation describing the arg.
These are things we can add in the future and update the format, keep
it simple for now.

Looking up config is based on naming convention, template file name to
config file. This is a little messy as we have config per variation,
rather than one canonical config per component, which contains variations,
or using collections. Thats being updated by Gemma on another branch, so
i'm ignoring that for now and updating the configs that make the components
exactly, which does the trick.

The args from config are used in the npm packaging step to rebuild the
macro structure. So the nunucks component output is the same (ignoring
whitespace).

There's quite a bit of duplication between to the transplier runners in
the template and component build step now. Its worth refactoring once
the component and template transpiling work has settled down a bit.

Small tweak to the preview setup, drop the Component suffix, avoid
relative paths. Use the dist version of templates, as the macros won't
exist in app/source anymore, and generally the styleguid should be
working on built, not source, files.