Load and configure Fractal using Gulp
Create a new Fractal instance and configure it for this project.

Amend the config file to add paths for the app root, components and
docs.

Add two new Gulp tasks to start the Fractal server and to run a static
export of the project web UI.

Install and load the Nunjucks adapter

https://github.com/frctl/nunjucks

- register the Nunjucks adapter for your components
- look for files with a .nunj file extension

Add the fractal task to the main Gulpfile

This will run the build task first, to ensure that any assets required
by the components are copied to /static.