Adds ERB components using meta-template
Use meta-template to transpile component templates

CAVEAT: This is using my GH branch which adds ERB support, via
package.json, which is better than having to use `npm link` but
not ideal long term.

You'll need to checkout meta-template, and that branch, and
from within this repo run `npm link ../meta-template` (or the
path of where you installed it)

Also deals with converting the default arugments (rails will
error if referenced when not set, and most arguments so far
are optional). Doesn't touch the main conversion work, which
we're hoping to do with meta-template.

Also wary of the manual erb construction going on, might be
better to add that as part of the nunjucks input before
the meta-template conversion, so the construction is only
in one language, not potetntially more