Add example assets for a component
I’m not sure we’ll want to use this, so I’ve only added it for one
component.

There must be an file in the component folder for the Asset tab to
appear. If there is more than one file, the Assets tab will have a
dropdown to select each asset.