Add a Gulp task to compile the Fractal theme stylesheet
This takes the theme stylesheet and copies it to a css folder in the
same directory.

For now, if the theme stylesheet can remain inside the theme folder,
rather than in /public, which is where the rest of the govuk frontend
alpha styles are - this seems like a sensible separation of the two.