Adds a basic phase-banner component implementation
This is a good example of adding a component from scratch, and was
easy to do. Basic component built in a couple of minutes, tested
against specs and a live-updating fractal. Worked well.

Used `message` as the param name, as that made more sense from an
API perspective, but doesn't match the BEM naming. Doesn't matter
in this case, I don't think, but would consistency there be useful?