Initialise component js by setting initScript in its config
This isn’t how we want to do this going forward, but I’d prefer to do
this in this way, rather than initialising all components in the
preview file.

This adds a if statement to see if context.setup.initScript exists, if
so,
the script for the component is included.