Fix the static build of Fractal
Manually copy our theme CSS and Images into the appropriate dirs.

I spoke to the Fractal team and they think this should happen automatically,
which makes sense for theme assets to be auto-copied. In the mean time
this allows us to use static builds if we want to.

Also ensures that the fractal theme assets have been compiled from source
before running the static build. This would've only worked if the fractal
server had run and the compiled assets happened to be there.

Note: The static build only work if the static output is served from `/`, not
directly from file system, or hosted subdir.