Add a test to ensure component args are defined
To avoid a new argument being added a component template, but not
added in the `arugments` config, which is used to generate dist
versions of components, eg adding `macro` syntax for nunjucks.

As it's an extra step it can easily be forgotten, which is what I
did when adding `messageHtml` to phase banner in early work.

This test ensures there are no arguments in variants that aren't
defined in the config, and will cause the test to fail if there are,
preventing that kind of bug from happening silently.