Make template file ext configurable
All our source templates use `njk` now, we should do the same for
output, but the output extension is tied to the name of the transpiler.

This convention works for a couple of transpilers, but it's not always
going to be a match, so allow it to be configured within the transpiler.

The `extensionForTarget` is pretty hacky, but felt cleaner than putting
the extension override in the build scripts, which would be duplicated
and it makes sense for each transpiler to be the authoraty on all things
for that format.

In the future i think we could take a more OOP approach to the transpilers,
with a base/abstract one, and each format extending on top. Then the build
tasks could instantiate the transpiler object it wants. This would
simplify the `transpileRunner` bits of build scripts.