Distribute nunjucks components in single file of macros
This feels like a nicer model for consuming apps, theres an example of
how it works on the node consumer app:

https://github.com/alphagov/govuk_frontend_alpha_npm_test/pull/10

This works from a consuming app point of view, but's this is tied to
the `npm` build, rather than the nunjucks format. We're not super
clear on the distinction bewteen the two at the moment.