Stop IE7 and lower from running component JS
We don’t test it, so it’s simpler to rely on progressive enhancement and stop IE7 and lower from accessing JS. This ties in with the removal of the JSON2 polyfill: it’s likely that code that relies on JSON would break leaving IE7 in a state where the `js-enabled` class had already been added.

We can leave the HTML5 shiv in place as it enables styles in older IE and doesn’t rely on JSON, but this renames the file reflect its contents.