Smoke test for style compilation
This moves the smoke test out of the components directory and turns it into a generic smoke test. The new styles compilation test just tries to compile the Sass and fails on an error.

Due to the need to run the compilation against the copied dist Sass, we need to run build:styles:copy first.