Fix the date component labels
I’d missed making the Month and Year labels programatically adjustable; this fixes that.

At the same time, I’ve added default values to the Nunjucks values. We’re currently using the config values for two purposes: as API documentation and as specific values for Fractal. It’s fine to have defaults in the config for Day/Month/Year to show how they could be customised, but it’s defensive design to also have defaults in the template here in case they aren’t provided.