Fix fractal theme not loading on non-root pages
This wasn't obviousy, as if you load the root page the path is fine,
and navigation within the app uses JS, so the assets stay loaded, but
if you refresh/load a non-root URL the path will fail.

Shame, because this makes it hard to run the static build 'in place'
and requires a web server to run.