Build a component script
Components need to own their own scripts. This moves the details polyfill over to the details component and updates the gulp task to build a components manifest as well.

With the details polyfill moved into the details component we can get rid of the existing elements manifest as it was then only including an old version of jQuery.

Finally, the template wasn’t including either elements or toolkit JS originally. Amended to add toolkit and the new components JS.