Watch component asset paths too
Watch was no longer catching SCSS/JS moved into the component
file structure, and wasn't getting rebuild/updating fractal.

The JS matcher works, but is too broad, and will trigger a
build if the config or spec JS is updated too. This doesn't
break anything, but it's probably not idea euther. Could
be refactored as with a more precice glob to exclude any
JS files with a double ext?