Customise fractal component tabs
Any view in mandlebrot is overrideable as a theme.

This demonstrates overriding the presence of a tab (hiding view), the name
of a tab (changing context to something else) and changing the content of
a tab (hijacking context to show component invocation).

Within the tab content/panel you still have access to the component object
so it's possible to generate dynamic content, eg the component invocation
is dynamic