Generate a tgz of the gem package
It's not possible to use the existing `.gem` package easily with a
rails app. While in alpha, and not shipping to rubygems, we want the
example consuming apps to have a managed reference to this package.

For npm we can point from package.json to a npm tgz hosted on github,
but you can't load a gem like that.

This approach generates a tgz of the gem, that can be commited to a repo,
as a binary file it won't show up in an editors but can be part of the
repo proper, not dependant on parent paths like the rails app is now.

The idea is the rails app will have a copy of the tgz committed, and
an additional build/install step which will untar into a gitignored
file, so that path can be used as a dep, but not be in git history
or a developers editor (assuming they have gitignore'd files hidden)