Allow ERB output in block default values
When transpiling ERB blocks with default values the default value was output
as an escaped/html_safe string, which mean it could not include any ERB logic.

For example:
```
{% block stylesheets %}<link href="{{ asset_path + 'file.css' }}" />{% endblock %}
```

Would become:
```
<% content_for?(:stylesheets) ? yield(:stylesheets) : '<link href="<%= asset_path \'file.css?1.0.0\'" \>.html_safe %>
```

Refactor how ERB output to support ERB syntax in the default clause. Outputting
the default directly will be the same as outputting it as an html_safe string. So
no change to any existing behaviour.