Improve ERB block-with-no-default transpile output
When transpiling a block with no default value to ERB don't use the
full if/else structure, dropping in and out of tags. Use the more
ruby idiomatic yield-if.

This isn't visible to ERB consumers in normal use, but if you're
inspecting the template output inside the gem, or if you're using
a debugging tool that gives you navigable stack traces, like
`better_errors` you can see whats going on in the template in a
cleaner way.