Fix broken asset path in govuk_template layout
The default stylesheet block's asset path is a malformed string,
which was causing the ERB transpiled output to blow up in a rails
consumer.

I'd have expected a syntax error would have blown up in nujucks too.
It's something we can add integration tests for in the future i think.