Automatically load Rails component helper in Gem
So a consuming rails app can call `govuk.button(...)` or similar
in a view without having to load the helper explicitly. Simplifies
the install process for a rails user.

Not sure this is a good long term approach, or that it's an idiomatic
rails/engine approach, but it works for alpha/testing. Should consult
a rubiest in the future and re-consider the technical implementation.

That said, not having to do anything as a consuming app, other than call
the helper feels like a nice experience, might be _too_ magic, but
hopefully that'll show up in testing.