Automatically load component macro in njk template
Simplies setup for component by not requiring an additional import for
components (`{% import "components.njk" as govuk %}`). In research we've
seen multiple people get tripped up by this.

Making it automatic will help with initial setup, but may make this too
magic and hard to work out where components come from (`govuk` won't be
referenced anywhere in the consuming app). We'll be conducting more user
research sessions with this change and that will give us an indication.

As part of this i've made it so the `npm` package step _only_ copies
templates in the `njk` format. This avoids having to conditionally prepend
the component import code for `njk` only, and it's the only format we're
testing with for the node build right now.