Only show the Introduction page for Components
This was also displaying when a subfolder was used within the docs.
Adding a check for item.isComponent ensures this is only shown for
items within the “components” directory.

Rename the Introduction section to Overview

This matches the convention used for the documentation, by default the
index page is called “Overview”.